+++ src/main/java/td/wallet/repository/CrudOperations.java
@@ -3,13 +3,13 @@ package td.wallet.repository;
 import java.util.List;
 
 public interface CrudOperations<T> {
-    T findById(Integer id);
+    T findById(long id);
 
     List<T> findAll();
 
     List<T> saveAll(List<T> toSave, List<Integer> relId);
 
-    T save(T toSave, int relId);
+    T save(T toSave, long relId);
 
     T delete(T toDelete);
 }
